Refactor work components: replace inline work rendering with WorksList and add WorksItem component
